+++ ChangeLog.md
@@ -14,6 +14,7 @@
     - services.pl workload-summary: fix errors management
     - TTP::execRemote() improve message readibility
     - services.pl commands: improve hash detection
+    - telemetry.pl ping: have the same help string whatever be the device
 
 ### 4.30.0